Separate token and grammar visitor;
